+++ android/src/main/java/com/aeyrium/sensor/AeyriumSensorPlugin.java
@@ -110,9 +110,8 @@ public class AeyriumSensorPlugin implements EventChannel.StreamHandler {
     float[] orientation = new float[3];
     SensorManager.getOrientation(adjustedRotationMatrix, orientation);
 
-    // Convert radians to degrees
-    double pitch = orientation[1]* -57;
-    double roll = orientation[2] * -57;
+    double pitch = - orientation[1];
+    double roll = - orientation[2];
     double[] sensorValues = new double[2];
     sensorValues[0] = pitch;
     sensorValues[1] = roll;