+++ lib/aeyrium_sensor.dart
@@ -6,11 +6,11 @@ const EventChannel _sensorEventChannel =
     EventChannel('plugins.aeyrium.com/sensor');
 
 class SensorEvent {
-  /// Pitch event from the device
+  /// Pitch from the device in radians
   /// A pitch is a rotation around a lateral (X) axis that passes through the device from side to side
   final double pitch;
 
-  ///Roll event from the device
+  ///Roll value from the device in radians
   ///A roll is a rotation around a longitudinal (Y) axis that passes through the device from its top to bottom
   final double roll;
 
@@ -21,20 +21,12 @@ class SensorEvent {
 }
 
 class AeyriumSensor {
-  static AeyriumSensor _instance;
-  Stream<SensorEvent> _sensorEvents;
+  static Stream<SensorEvent> _sensorEvents;
 
-  factory AeyriumSensor() {
-    if (_instance == null) {
-      _instance = AeyriumSensor._internal();
-    }
-    return _instance;
-  }
-
-  AeyriumSensor._internal();
+  AeyriumSensor._();
 
   /// A broadcast stream of events from the device rotation sensor.
-  Stream<SensorEvent> get sensorEvents {
+  static Stream<SensorEvent> get sensorEvents {
     if (_sensorEvents == null) {
       _sensorEvents = _sensorEventChannel
           .receiveBroadcastStream()
@@ -43,7 +35,7 @@ class AeyriumSensor {
     return _sensorEvents;
   }
 
-  SensorEvent _listToSensorEvent(List<double> list) {
+  static SensorEvent _listToSensorEvent(List<double> list) {
     return SensorEvent(list[0], list[1]);
   }
 }