+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.4] -  10/02/2018
+
+* Values are expressed in radians instead of degrees
+* Use AeyriumSensor.sensorEvents instead of AeyriumSensor().sensorEvents
+
 ## [1.0.3] -  10/02/2018
 
 * Pitch value was fixed when rotate the device on iOS